Confine the float oracle sweeps to the CMSIS backend's 32..4096 on the M55
The first QEMU run of the rebased suite HardFaulted in
fft_oracle_test/0.ImpulseIsFlatAtEverySize at N = 4 on the cortex-m55 leg,
the one leg where basic_real_fft<float> routes through CMSIS-DSP:
arm_rfft_fast_init_f32 accepts only 32..4096 and fft.h's wrapper does not
check its return status, so fft.h's ">= 4" contract is undefined behaviour
there below 32 (the existing battery starts at 64, which is why it never
showed). The oracle's profile<Sample> now carries the engine's size range
(k_min_n / k_max_n; 32..4096 for float under TAP_DSP_FFT_CMSIS, 4..2^20
otherwise) and every sweep is built from it. The wrapper defect itself is
recorded for the fft.h owner, not papered over here.

Also: the informational table printed "zu" for N on newlib-nano; %lu now.

Hosted GCC 13 / Clang 18: 214/214, -Werror clean; host one-shot mode:
three markers rc=0, selected=161/4/1, skipped=0.

+++ tests/test_fft_parity_ooura.cpp
@@ -401,8 +401,10 @@ namespace {
                 }
             }
             overall = std::max({overall, fwd, inv});
-            std::printf("[ ulp ] %10zu %14llu %14llu   %s / %s\n", n, static_cast<unsigned long long>(fwd),
-                        static_cast<unsigned long long>(inv), fwd_worst, inv_worst);
+            // %lu, not %zu: newlib-nano's printf on the QEMU legs prints "zu".
+            std::printf("[ ulp ] %10lu %14llu %14llu   %s / %s\n", static_cast<unsigned long>(n),
+                        static_cast<unsigned long long>(fwd), static_cast<unsigned long long>(inv), fwd_worst,
+                        inv_worst);
             ::testing::Test::RecordProperty(std::string(precision_name<Sample>()) + "_N" + std::to_string(n) + "_fwd",
                                             std::to_string(fwd));
             ::testing::Test::RecordProperty(std::string(precision_name<Sample>()) + "_N" + std::to_string(n) + "_inv",
